requirements.txt
asgiref==3.5.2
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.2
django-user-agents==0.4.0
idna==2.10
iniconfig==1.1.1
jwt==1.2.0
oauthlib==3.1.0
packaging==21.0
Pillow==9.2.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.1
python-dotenv==0.20.0
python3-openid==3.2.0
pytube @ git+https://github.com/pytube/pytube@2e307d8d615ef30aa837fe2275954146dab07ca6
pytz==2020.4
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
toml==0.10.2
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
youtube-dl==2021.6.6
